# api-gateway/swagger.yaml
openapi: 3.0.3
info:
  title: CommerceAI Pro API
  description: |
    API Gateway pour la plateforme CommerceAI Pro - Intelligence Artificielle pour E-commerce
    
    Cette API fournit un accès complet aux fonctionnalités de la plateforme :
    - Authentification et gestion des utilisateurs
    - Gestion des organisations et rôles
    - Support client et chat en temps réel
    - Analytics et métriques
    - Intégrations avec les plateformes e-commerce
    - Agents IA spécialisés
    
    ## Authentification
    
    L'API utilise l'authentification JWT. Incluez le token dans l'en-tête Authorization :
    ```
    Authorization: Bearer <votre-token-jwt>
    ```
    
    ## Codes de statut
    
    - `200` - Succès
    - `201` - Créé avec succès
    - `400` - Erreur de validation
    - `401` - Non authentifié
    - `403` - Accès refusé
    - `404` - Ressource non trouvée
    - `429` - Limite de taux dépassée
    - `500` - Erreur serveur
    
  version: 1.0.0
  contact:
    name: Support CommerceAI Pro
    email: support@commerceai-pro.com
    url: https://commerceai-pro.com
  license:
    name: Propriétaire
    url: https://commerceai-pro.com/license
servers:
  - url: https://api.commerceai-pro.com
    description: Production
  - url: https://staging-api.commerceai-pro.com
    description: Staging
  - url: http://localhost:4000
    description: Développement local
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Erreur lors de l'opération"
        error:
          type: string
          example: "Détails de l'erreur"
        timestamp:
          type: string
          format: date-time
    Success:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Opération réussie"
        data:
          type: object
        timestamp:
          type: string
          format: date-time
...
    Organization:
      type: object
      properties:
        id:
          type: string
          example: "org_123"
        name:
          type: string
          example: "Mon Entreprise"
        domain:
          type: string
          example: "monentreprise.com"
        plan:
          type: string
          enum: [starter, professional, enterprise]
          example: "professional"
        settings:
          type: object
        createdAt:
          type: string
          format: date-time
...
security:
  - BearerAuth: []
paths:
...
  /api/organizations:
    get:
      tags:
        - Organisations
      summary: Liste des organisations
      description: Récupérer la liste des organisations (admin uniquement)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Liste des organisations
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Organization'
        '403':
          description: Accès refusé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/organizations/{orgId}:
    get:
      tags:
        - Organisations
      summary: Détails d'une organisation
      description: Récupérer les détails d'une organisation
      parameters:
        - name: orgId
          in: path
          required: true
          description: ID de l'organisation
          schema:
            type: string
      responses:
        '200':
          description: Détails de l'organisation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Organization'
        '404':
          description: Organisation non trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  # Analytics
...